fixed data dependency
Previously I was using the imported Data as reference on my components instead of referencing this.state. Refactored code to reference this.state.roverData for the data object instead. This will allow me to change my data source on the fly which is required to change what is rendered depending on the rover selected.